refactor: 패널 정리
