+++ package.json
@@ -16,6 +16,7 @@
   "dependencies": {
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
+    "axios": "^1.7.9",
     "jotai": "^2.11.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
